Build the diag harnesses without warnings
- printf("%d") received qsizetype (64 bit) in 21 places; the counts are
  now passed as int, as the other harnesses already do.
- Six QFile::open results were ignored. A failed open would have left a
  harness reading an empty file or testing preview cleanup without its
  decoys, and reporting a misleading result. test_window_geometry checks
  the open like any other expectation, test_preview_clip fails when it
  cannot create the decoys, test_container_sync stops when it cannot
  write the test configuration.

+++ tools/diag/test_pillarbox.cpp
@@ -125,7 +125,7 @@ int main(int argc, char** argv)
 
     const QList<TTFrameInfo>& idx = w.frameIndex();
     const TTDisplayOrderMap&  map = w.displayOrderMap();
-    printf("frames=%d gops=%d\n", idx.size(), (int)bundle.gops.size());
+    printf("frames=%d gops=%d\n", (int)idx.size(), (int)bundle.gops.size());
 
     int nI = 0;
     for (const auto& fi : idx)
